requirements.txt
altgraph==0.17.3
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==38.0.3
docx2txt==0.8
EbookLib==0.17.1
future==0.18.2
html2text==2020.1.16
idna==3.4
loguru==0.6.0
lxml==4.9.1
mobi==0.3.3
pdfminer.six==20221105
pefile==2022.5.30
Pillow==9.3.0
psutil==5.9.4
pycparser==2.21
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQt5-stubs==5.15.6.0
python-pptx==0.6.21
pywin10==0.0.3
pywin32==305
pywin32-ctypes==0.2.0
requests==2.28.1
six==1.16.0
sqlitedict==2.0.0
urllib3==1.26.13
win32-setctime==1.1.0
winshell==0.6
xlrd==2.0.1
XlsxWriter==3.0.3
